backend/pyproject.toml
[project]
name = "merry-backend"
version = "1.0.0"
description = "Merry - AI Document Generation Platform"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.100.0",
    "uvicorn>=0.23.0",
    "pydantic>=2.0.0",
    "python-dotenv>=1.0.0",
    "sqlalchemy[asyncio]>=2.0.0",
    "asyncpg>=0.28.0",
    "supabase>=2.0.0",
    "groq>=0.4.0",
    "httpx>=0.28.0",
    "python-docx>=0.8.11",
    "openpyxl>=3.1.0",
    "email-validator>=2.0.0",
]

[project.optional-dependencies]
dev = ["pytest>=8.0.0", "pytest-asyncio>=0.23.0"]
